+++ src/expression_evaluator/lambda_evaluator.cpp
@@ -3,6 +3,7 @@
 #include "binder/expression/lambda_expression.h"
 #include "binder/expression/scalar_function_expression.h"
 #include "common/exception/runtime.h"
+#include "common/system_config.h"
 #include "expression_evaluator/expression_evaluator_visitor.h"
 #include "function/list/vector_list_functions.h"
 #include "parser/expression/parsed_lambda_expression.h"
