+++ src/graph/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_library(kuzu_graph
         OBJECT
+        graph.cpp
         graph_entry.cpp
         on_disk_graph.cpp)
