+++ src/include/expression_evaluator/case_evaluator.h
@@ -3,6 +3,7 @@
 #include <bitset>
 
 #include "binder/expression/expression.h"
+#include "common/system_config.h"
 #include "expression_evaluator.h"
 
 namespace kuzu {
