+++ src/processor/operator/unwind.cpp
@@ -1,5 +1,6 @@
 #include "processor/operator/unwind.h"
 
+#include "common/system_config.h"
 #include "processor/execution_context.h"
 
 using namespace kuzu::common;
