+++ test/storage/node_insertion_deletion_test.cpp
@@ -1,4 +1,5 @@
 #include "common/constants.h"
+#include "common/system_config.h"
 #include "graph_test/graph_test.h"
 
 using namespace kuzu::common;
